+++ src/sparsezoo/models/__init__.py
@@ -3,3 +3,4 @@ Convenient model constructors for common models in the sparsezoo
 """
 
 # flake8: noqa
+from .zoo import *
