+++ src/sparsezoo/models/classification/inception.py
@@ -5,8 +5,8 @@ Inception models:
 
 from typing import Union
 
+from sparsezoo.models.zoo import Zoo
 from sparsezoo.objects import Model
-from sparsezoo.zoo import load_model
 
 
 __all__ = ["inception_v3"]
@@ -49,7 +49,7 @@ def inception_v3(
     :param force_token_refresh: True to refresh the auth token, False otherwise
     :return: The created model
     """
-    return load_model(
+    return Zoo.load_model(
         domain="cv",
         sub_domain="classification",
         architecture="inception_v3",
