+++ src/sparsezoo/models/classification/mobilenet.py
@@ -6,8 +6,8 @@ MobileNet models:
 
 from typing import Union
 
+from sparsezoo.models.zoo import Zoo
 from sparsezoo.objects import Model
-from sparsezoo.zoo import load_model
 
 
 __all__ = ["mobilenet_v1", "mobilenet_v2"]
@@ -50,7 +50,7 @@ def mobilenet_v1(
     :param force_token_refresh: True to refresh the auth token, False otherwise
     :return: The created model
     """
-    return load_model(
+    return Zoo.load_model(
         domain="cv",
         sub_domain="classification",
         architecture="mobilenet_v1",
@@ -105,7 +105,7 @@ def mobilenet_v2(
     :param force_token_refresh: True to refresh the auth token, False otherwise
     :return: The created model
     """
-    return load_model(
+    return Zoo.load_model(
         domain="cv",
         sub_domain="classification",
         architecture="mobilenet_v2",
