+++ src/sparsezoo/models/classification/vgg.py
@@ -5,8 +5,8 @@ VGG models:
 
 from typing import Union
 
+from sparsezoo.models.zoo import Zoo
 from sparsezoo.objects import Model
-from sparsezoo.zoo import load_model
 
 
 __all__ = [
@@ -58,7 +58,7 @@ def vgg_11(
     :param force_token_refresh: True to refresh the auth token, False otherwise
     :return: The created model
     """
-    return load_model(
+    return Zoo.load_model(
         domain="cv",
         sub_domain="classification",
         architecture="vgg",
@@ -113,7 +113,7 @@ def vgg_11bn(
     :param force_token_refresh: True to refresh the auth token, False otherwise
     :return: The created model
     """
-    return load_model(
+    return Zoo.load_model(
         domain="cv",
         sub_domain="classification",
         architecture="vgg",
@@ -168,7 +168,7 @@ def vgg_13(
     :param force_token_refresh: True to refresh the auth token, False otherwise
     :return: The created model
     """
-    return load_model(
+    return Zoo.load_model(
         domain="cv",
         sub_domain="classification",
         architecture="vgg",
@@ -223,7 +223,7 @@ def vgg_13bn(
     :param force_token_refresh: True to refresh the auth token, False otherwise
     :return: The created model
     """
-    return load_model(
+    return Zoo.load_model(
         domain="cv",
         sub_domain="classification",
         architecture="vgg",
@@ -278,7 +278,7 @@ def vgg_16(
     :param force_token_refresh: True to refresh the auth token, False otherwise
     :return: The created model
     """
-    return load_model(
+    return Zoo.load_model(
         domain="cv",
         sub_domain="classification",
         architecture="vgg",
@@ -333,7 +333,7 @@ def vgg_16bn(
     :param force_token_refresh: True to refresh the auth token, False otherwise
     :return: The created model
     """
-    return load_model(
+    return Zoo.load_model(
         domain="cv",
         sub_domain="classification",
         architecture="vgg",
@@ -388,7 +388,7 @@ def vgg_19(
     :param force_token_refresh: True to refresh the auth token, False otherwise
     :return: The created model
     """
-    return load_model(
+    return Zoo.load_model(
         domain="cv",
         sub_domain="classification",
         architecture="vgg",
@@ -443,7 +443,7 @@ def vgg_19bn(
     :param force_token_refresh: True to refresh the auth token, False otherwise
     :return: The created model
     """
-    return load_model(
+    return Zoo.load_model(
         domain="cv",
         sub_domain="classification",
         architecture="vgg",
